make thumbnails a little narrower
@jerrodanzalone, adjust these widths if you'd like.

DUR figure: even at 100%, it was too small to see without zooming in.  So I made the thumbnail smaller so it didn't consume as much real estate.

Team member figure: I made it smaller because it was plenty readable on my screen.  Why am I a checkbox in that figure?  Feels impersonal.  And why are we all viruses?  Yes I know some of us have cooties --but all of us?  And we have to advertise it?